apps/provider_analystco/requirements.txt
fastapi
uvicorn[standard]
whoosh
markdown2
pydantic
python-slugify
pyyaml